requirements.txt
asgiref==3.5.2
certifi @ file:///croot/certifi_1665076670883/work/certifi
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.1.3
django-filter==22.1
django-phonenumber-field==7.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
Markdown==3.4.1
phonenumberslite==8.13.1
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.13
